реализация тренадцатого инкремента

+++ migrations/000001_create_metrics_table.up.sql
@@ -3,5 +3,5 @@ CREATE TABLE metrics (
     id VARCHAR(255) PRIMARY KEY,
     mtype VARCHAR(255) NOT NULL,
     val DOUBLE PRECISION,
-    delta INTEGER
+    delta BIGINT
 );
